Addressed PR feedback - connect function changes, IsServerless function, no use of GoCBv2 in Rest, existing db specific code commented out, new file for serverless tests
