+++ rest/server_context.go
@@ -826,7 +826,7 @@ func dbcOptionsFromConfig(sc *ServerContext, config *DbConfig, dbName string) (d
 		BcryptCost:                bcryptCost,
 		GroupID:                   groupID,
 		JavascriptTimeout:         javascriptTimeout,
-		Serverless:                base.BoolDefault(sc.config.Unsupported.Serverless, false),
+		Serverless:                sc.config.IsServerless(),
 	}
 
 	return contextOptions, nil
